+++ CHANGELOG.md
@@ -77,6 +77,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ServiceUnavailableException` added
 - `Project` model added
 - `project_object` added to `Observation` model
+- `id_observation` added to `Detail` model
 
 ### Changed